+++ lib/bh/railtie.rb
@@ -11,6 +11,7 @@ require 'bh/helpers/nav_helper'
 require 'bh/helpers/navbar_helper'
 require 'bh/helpers/panel_helper'
 require 'bh/helpers/panel_row_helper'
+require 'bh/helpers/progress_bar_helper'
 
 module Bh
   class Railtie < Rails::Railtie
@@ -28,6 +29,7 @@ module Bh
       ActionView::Base.send :include, NavbarHelper
       ActionView::Base.send :include, PanelHelper
       ActionView::Base.send :include, PanelRowHelper
+      ActionView::Base.send :include, ProgressBarHelper
     end
 
     initializer 'bh.add_views' do |app|